chore: remove pkg.lock.json and rename trainingSession file
