+++ components/programBuilder/TrainingSessionExercise.tsx
